requirements.txt
requests
websocket-client
beautifulsoup4
pillow